+++ source/App/source/WebApp/WebApp.csproj
@@ -63,6 +63,10 @@ limitations under the License.
   </PropertyGroup>
 
   <ItemGroup>
+    <InternalsVisibleTo Include="ExampleHost.WebApi04" />
+  </ItemGroup>
+
+  <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
